requirements.txt
django==2.2.17
djangorestframework==3.12.2
requests